# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Portal Marujo - CSA Statistics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /managers/{id}:
    get:
      operationId: getManager
      tags: [managers]
      summary: Get manager statistics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Manager details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Season:
      type: object
      required: [year, matches, wins, draws, losses, goalsScored, goalsConceded]
      properties:
        year:
          type: string
        matches:
          type: integer
        wins:
          type: integer
        draws:
          type: integer
        losses:
          type: integer
        goalsScored:
          type: integer
        goalsConceded:
          type: integer
        topScorer:
          type: ["string", "null"]
        topScorerGoals:
          type: ["integer", "null"]
        topAppearances:
          type: ["string", "null"]
        topAppearancesCount:
          type: ["integer", "null"]
        statsFullyVerified:
          type: boolean
          description: Season sheets/stats marked fully verified
        statsVerifiedAt:
          type: ["string", "null"]
          format: date-time
...
    ManagerDetail:
      type: object
      required: [id, name, matches, wins, draws, losses, winPercentage, seasonStats]
      properties:
        id:
          type: integer
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        startYear:
          type: ["integer", "null"]
        endYear:
          type: ["integer", "null"]
        verificationStatus:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
          format: date-time
        verifiedBy:
          type: ["string", "null"]
        matches:
          type: integer
        wins:
          type: integer
        draws:
          type: integer
        losses:
          type: integer
        winPercentage:
          type: number
        goalsScored:
          type: integer
        goalsConceded:
          type: integer
        seasonStats:
          type: array
          items:
            $ref: "#/components/schemas/Season"
...